+++ crates/dss-core/tests/golden_flicker.rs
@@ -50,8 +50,9 @@ fn flicker_meter_matches_r3723_bit_exact() {
 
     // Pascal `Vbase: Single := 1000*kVBase` (narrowed to f32 before use).
     let vbase = ((1000.0 * g.kvbase) as f32) as f64;
-    // Npst = 1 + Trunc(t_N / 600) (Monitor.pas:1651).
-    let npst = 1 + (g.times[g.n - 1] / 600.0).trunc() as usize;
+    // Npst = 1 + Trunc(t_N / 600) (Monitor.pas:1651); the division is Double
+    // (Single numerator promoted, Double literal), mirroring `post.rs`.
+    let npst = 1 + (g.times[g.n - 1] as f64 / 600.0).trunc() as usize;
 
     for p in 0..g.nphases {
         let mut prms = g.raw_mag[p].clone();
@@ -72,7 +73,9 @@ fn flicker_meter_matches_r3723_bit_exact() {
         let mut t_pst = 0.0_f32;
         let mut ipst = 0usize;
         for i in 0..g.n {
-            if (g.times[i] - t_pst) >= 600.0 {
+            // Single operands promoted to Double for the compare (Delphi Win64),
+            // mirroring `post.rs::do_flicker_calculations`.
+            if (g.times[i] as f64 - t_pst as f64) >= 600.0 {
                 ipst += 1;
                 t_pst = g.times[i];
             }
